split off TestDatetimeIndex class
